Notify when ws goes offline
