+++ src/model/connection-status.ts
@@ -1,0 +1,7 @@
+export enum ConnectionStatus {
+  DISCONNECTED = 'DISCONNECTED',
+  CONNECTING = 'CONNECTING',
+  INITIALIZING = 'INITIALIZING',
+  CONNECTED = 'CONNECTED',
+  RECONNECTING = 'RECONNECTING',
+}
